chatbot/requirements.txt
altgraph==0.17
asgiref==3.2.10
auto-py-to-exe==2.7.8
bottle==0.12.18
bottle-websocket==0.2.9
certifi==2020.6.20
chardet==3.0.4
dbconfig==0.3.2
Django==3.1.2
Eel==0.12.4
future==0.18.2
gevent==20.9.0
gevent-websocket==0.10.1
greenlet==0.4.17
idna==2.10
Pillow==7.2.0
pony==0.7.13
psycopg2-binary==2.8.6
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.9
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.4.0
urllib3==1.25.10
vk-api==11.9.0
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.1.2
